requirements.txt
astroid==1.6.6
backports.functools-lru-cache==1.5
configparser==4.0.2
confusable-homoglyphs==3.2.0
Django==1.11.25
django-bootstrap3==11.1.0
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.9.4
enum34==1.1.6
futures==3.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pylint==1.9.5
python-decouple==3.1
pytz==2019.3
singledispatch==3.4.0.3
six==1.12.0
whitenoise==4.1.4
wrapt==1.11.2
